Crate w_result

Defines WResult, a result type that carries warnings.

Sometimes it may be possible for an operation to proceed despite encountering errors. In these cases, the caller may need to know about the errors that occured. WResult is similar to Result except that the ok variant carries a vector of accumulated warnings. It comes with methods for converting to a Result by discarding or logging the warnings or treating them as errors.

Reexports

pub use self::WResult::*;

Enums

WResult
A result type that carries warnings.